requirements.txt
antlr4-python3-runtime==4.9.3
appnope==0.1.4
asttokens==3.0.0
botorch==0.13.0
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
filelock==3.18.0
fonttools==4.57.0
fsspec==2025.3.2
gpytorch==1.14
hydra-core==1.3.2
ipykernel==6.29.5
ipython<9,>=8.17.2
ipython_pygments_lexers==1.1.1
jaxtyping==0.3.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
linear-operator==0.6
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mpmath==1.3.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
omegaconf==2.3.0
opt_einsum==3.4.0
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
pyre-extensions==0.0.32
pyro-api==0.1.2
pyro-ppl==1.9.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyzmq==26.4.0
scikit-learn==1.6.1
scipy==1.15.2
setuptools==78.1.0
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.6.0
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.13.1
wadler_lindig==0.1.4
wcwidth==0.2.13
hydra-submitit-launcher==1.2.0